# docs/openapi copy.yaml
openapi: 3.1.0
info:
  title: Star Wars API
  description: API for interacting with Star Wars data.
  version: 1.0.0
paths:
...
  /planets/{id}:
    get:
      summary: Get a planet by ID
      description: Retrieve details about a specific planet by its ID.
      operationId: getPlanetById
      parameters:
        - name: id
          in: path
          description: The ID of the planet to retrieve.
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Details of the planet.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "ok"
                  result:
                    $ref: "#/components/schemas/PlanetDetail"
components:
  schemas:
...
    PlanetDetail:
      type: object
      properties:
        properties:
          type: object
          properties:
            diameter:
              type: string
              example: "10465"
            rotation_period:
              type: string
              example: "23"
            orbital_period:
              type: string
              example: "304"
            gravity:
              type: string
              example: "1 standard"
            population:
              type: string
              example: "200000"
            climate:
              type: string
              example: "arid"
            terrain:
              type: string
              example: "desert"
            surface_water:
              type: string
              example: "1"
            created:
              type: string
              format: date-time
              example: "2024-12-17T15:23:44.809Z"
            edited:
              type: string
              format: date-time
              example: "2024-12-17T15:23:44.809Z"
            name:
              type: string
              example: "Tatooine"
            url:
              type: string
              example: "https://www.swapi.tech/api/planets/1"
        description:
          type: string
          example: "A planet."
        _id:
          type: string
          example: "5f7254c11b7dfa00041c6fae"
        uid:
          type: string
          example: "1"
        __v:
          type: integer
          example: 0